# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Production-ready AI Agent API with DDD architecture and Eino workflow engine. Supports OpenAI-compatible tool calling, human-in-the-loop approval, JWT authentication, and streaming responses.",
        "title": "Go Agent API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "https://github.com/wyuneed/go-agent-api/issues"
        },
        "license": {
            "name": "MIT",
            "url": "https://github.com/wyuneed/go-agent-api/blob/main/LICENSE"
        },
        "version": "1.0"
    },
    "host": "127.0.0.1:8080",
    "basePath": "/",
    "paths": {
...
        "/v1/chat/completions": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Sends messages to the AI agent and returns a completion. Set stream=true to receive Server-Sent Events.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "chat"
                ],
                "summary": "Chat completions (OpenAI-compatible)",
                "parameters": [
                    {
                        "description": "Chat completion request",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_wyuneed_go-agent-api_internal_application_dto_request.ChatCompletionRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_wyuneed_go-agent-api_internal_pkg_response.SwaggerChatCompletionResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_wyuneed_go-agent-api_internal_pkg_response.SwaggerErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_wyuneed_go-agent-api_internal_pkg_response.SwaggerErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_wyuneed_go-agent-api_internal_pkg_response.SwaggerErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1/conversations": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns all conversations belonging to the authenticated user",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "conversations"
                ],
                "summary": "List conversations",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_wyuneed_go-agent-api_internal_pkg_response.SwaggerConversationListResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_wyuneed_go-agent-api_internal_pkg_response.SwaggerErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_wyuneed_go-agent-api_internal_pkg_response.SwaggerErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Creates a new conversation and sends the first message to the agent",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "conversations"
                ],
                "summary": "Create a conversation",
                "parameters": [
                    {
                        "description": "Conversation details",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_wyuneed_go-agent-api_internal_application_dto_request.CreateConversationRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/github_com_wyuneed_go-agent-api_internal_pkg_response.SwaggerSendMessageResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_wyuneed_go-agent-api_internal_pkg_response.SwaggerErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_wyuneed_go-agent-api_internal_pkg_response.SwaggerErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_wyuneed_go-agent-api_internal_pkg_response.SwaggerErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1/conversations/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns a conversation with its full message history",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "conversations"
                ],
                "summary": "Get a conversation",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Conversation ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_wyuneed_go-agent-api_internal_pkg_response.SwaggerConversationResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_wyuneed_go-agent-api_internal_pkg_response.SwaggerErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_wyuneed_go-agent-api_internal_pkg_response.SwaggerErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_wyuneed_go-agent-api_internal_pkg_response.SwaggerErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1/conversations/{id}/approve": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Resumes a conversation that is paused waiting for human-in-the-loop approval",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "conversations"
                ],
                "summary": "Approve or reject a pending action",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Conversation ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Approval decision",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_wyuneed_go-agent-api_internal_application_dto_request.ApprovalRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_wyuneed_go-agent-api_internal_pkg_response.SwaggerApprovalResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_wyuneed_go-agent-api_internal_pkg_response.SwaggerErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_wyuneed_go-agent-api_internal_pkg_response.SwaggerErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_wyuneed_go-agent-api_internal_pkg_response.SwaggerErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1/conversations/{id}/messages": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Sends a message to an existing conversation and returns the agent's reply",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "conversations"
                ],
                "summary": "Send a message",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Conversation ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Message content",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_wyuneed_go-agent-api_internal_application_dto_request.SendMessageRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_wyuneed_go-agent-api_internal_pkg_response.SwaggerSendMessageResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_wyuneed_go-agent-api_internal_pkg_response.SwaggerErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_wyuneed_go-agent-api_internal_pkg_response.SwaggerErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_wyuneed_go-agent-api_internal_pkg_response.SwaggerErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "github_com_wyuneed_go-agent-api_internal_application_dto_request.ApprovalRequest": {
            "type": "object",
            "properties": {
                "approved": {
                    "type": "boolean"
                },
                "reason": {
                    "type": "string"
                }
            }
        },
        "github_com_wyuneed_go-agent-api_internal_application_dto_request.ChatCompletionRequest": {
            "type": "object",
            "properties": {
                "max_tokens": {
                    "type": "integer"
                },
                "messages": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_wyuneed_go-agent-api_internal_application_port.ChatMessage"
                    }
                },
                "model": {
                    "type": "string"
                },
                "stream": {
                    "type": "boolean"
                },
                "temperature": {
                    "type": "number"
                },
                "tool_choice": {},
                "tools": {
                    "type": "array",
                    "items": {}
                }
            }
        },
        "github_com_wyuneed_go-agent-api_internal_application_dto_request.CreateConversationRequest": {
            "type": "object",
            "properties": {
                "agent_type": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "github_com_wyuneed_go-agent-api_internal_application_dto_request.SendMessageRequest": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
        "github_com_wyuneed_go-agent-api_internal_application_port.ChatMessage": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "role": {
                    "type": "string"
                },
                "tool_call_id": {
                    "type": "string"
                },
                "tool_calls": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_wyuneed_go-agent-api_pkg_toolspec.ToolCall"
                    }
                }
            }
        },
        "github_com_wyuneed_go-agent-api_internal_pkg_response.SwaggerApprovalData": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Action approved, resuming workflow"
                },
                "status": {
                    "type": "string",
                    "example": "approved"
                }
            }
        },
        "github_com_wyuneed_go-agent-api_internal_pkg_response.SwaggerApprovalResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/github_com_wyuneed_go-agent-api_internal_pkg_response.SwaggerApprovalData"
                },
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
...
        "github_com_wyuneed_go-agent-api_internal_pkg_response.SwaggerChatCompletionChoice": {
            "type": "object",
            "properties": {
                "finish_reason": {
                    "type": "string",
                    "example": "stop"
                },
                "index": {
                    "type": "integer",
                    "example": 0
                },
                "message": {
                    "$ref": "#/definitions/github_com_wyuneed_go-agent-api_internal_pkg_response.SwaggerMessageData"
                }
            }
        },
        "github_com_wyuneed_go-agent-api_internal_pkg_response.SwaggerChatCompletionData": {
            "type": "object",
            "properties": {
                "choices": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_wyuneed_go-agent-api_internal_pkg_response.SwaggerChatCompletionChoice"
                    }
                },
                "id": {
                    "type": "string",
                    "example": "chatcmpl-abc123"
                },
                "model": {
                    "type": "string",
                    "example": "gpt-4o-mini"
                },
                "object": {
                    "type": "string",
                    "example": "chat.completion"
                }
            }
        },
        "github_com_wyuneed_go-agent-api_internal_pkg_response.SwaggerChatCompletionResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/github_com_wyuneed_go-agent-api_internal_pkg_response.SwaggerChatCompletionData"
                },
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
        "github_com_wyuneed_go-agent-api_internal_pkg_response.SwaggerConversationData": {
            "type": "object",
            "properties": {
                "agent_type": {
                    "type": "string",
                    "example": "general"
                },
                "created_at": {
                    "type": "string",
                    "example": "2024-01-01T00:00:00Z"
                },
                "id": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                },
                "message_count": {
                    "type": "integer",
                    "example": 5
                },
                "model": {
                    "type": "string",
                    "example": "gpt-4o-mini"
                },
                "status": {
                    "type": "string",
                    "example": "active"
                },
                "title": {
                    "type": "string",
                    "example": "My first conversation"
                },
                "total_tokens": {
                    "type": "integer",
                    "example": 1024
                },
                "updated_at": {
                    "type": "string",
                    "example": "2024-01-01T00:01:00Z"
                },
                "user_id": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                }
            }
        },
        "github_com_wyuneed_go-agent-api_internal_pkg_response.SwaggerConversationListResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_wyuneed_go-agent-api_internal_pkg_response.SwaggerConversationData"
                    }
                },
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
        "github_com_wyuneed_go-agent-api_internal_pkg_response.SwaggerConversationResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/github_com_wyuneed_go-agent-api_internal_pkg_response.SwaggerConversationData"
                },
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
        "github_com_wyuneed_go-agent-api_internal_pkg_response.SwaggerErr": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string",
                    "example": "INVALID_REQUEST"
                },
                "message": {
                    "type": "string",
                    "example": "Invalid request body"
                }
            }
        },
        "github_com_wyuneed_go-agent-api_internal_pkg_response.SwaggerErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "$ref": "#/definitions/github_com_wyuneed_go-agent-api_internal_pkg_response.SwaggerErr"
                },
                "success": {
                    "type": "boolean",
                    "example": false
                }
            }
        },
...
        "github_com_wyuneed_go-agent-api_internal_pkg_response.SwaggerMessageData": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string",
                    "example": "Hello! How can I help you today?"
                },
                "conversation_id": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                },
                "created_at": {
                    "type": "string",
                    "example": "2024-01-01T00:00:00Z"
                },
                "id": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                },
                "model": {
                    "type": "string",
                    "example": "gpt-4o-mini"
                },
                "role": {
                    "type": "string",
                    "example": "assistant"
                }
            }
        },
...
        "github_com_wyuneed_go-agent-api_internal_pkg_response.SwaggerSendMessageData": {
            "type": "object",
            "properties": {
                "conversation": {
                    "$ref": "#/definitions/github_com_wyuneed_go-agent-api_internal_pkg_response.SwaggerConversationData"
                },
                "message": {
                    "$ref": "#/definitions/github_com_wyuneed_go-agent-api_internal_pkg_response.SwaggerMessageData"
                }
            }
        },
        "github_com_wyuneed_go-agent-api_internal_pkg_response.SwaggerSendMessageResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/github_com_wyuneed_go-agent-api_internal_pkg_response.SwaggerSendMessageData"
                },
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
...
        "github_com_wyuneed_go-agent-api_pkg_toolspec.FunctionCall": {
            "type": "object",
            "properties": {
                "arguments": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "github_com_wyuneed_go-agent-api_pkg_toolspec.ToolCall": {
            "type": "object",
            "properties": {
                "function": {
                    "$ref": "#/definitions/github_com_wyuneed_go-agent-api_pkg_toolspec.FunctionCall"
                },
                "id": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Enter \"Bearer\" followed by a space and your JWT token. Example: \"Bearer eyJhbGci...\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}